+++ src/main/resources/templates/index.html
@@ -7,4 +7,4 @@
 <body>
 Hello World!!
 </body>
-</html
+</html>
